Riorganizzazione package
