--- src/model/LoginBean.java
@@ -1,59 +1,0 @@
-package model;
-
-
-public class LoginBean {
-	private String username;
-	private String password;
-	private String nome;
-	private String cognome;
-
-	public LoginBean() {
-		this.username = "";
-		this.password = "";
-		this.nome = "";
-	}
-
-	public void setUsername(String user) {
-		this.username = user;
-	}
-
-	public String getUsername() {
-		return this.username;
-	}
-
-	public void setPassword(String pwd) {
-		this.password = pwd;
-	}
-
-	public String getPassword() {
-		return this.password;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public String getNome() {
-		return this.nome;
-	}
-
-	public void setCognome(String cogn) {
-		this.cognome = cogn;
-	}
-
-	public String getCognome() {
-		return this.cognome;
-	}
-
-	public boolean validate() {
-		// Controllo sintattico
-		System.out.println("prima del metodo validate");
-		if (this.username.equals("") || this.password.equals("")) {
-			return false;
-		}
-		LoginController controller = LoginController.getInstance();
-		Usr found = controller.login(this.username, this.password);
-		return  (found != null);
-	}
-
-}
